Add pyright language server configuration
The nvim-lspconfig plugin configuration is loaded after the lsp/*
files in the nvim rtp. This results in the on_attach as is provided
by the configuration in that plugin. Since we want to detect
where python lives (because of venv, or uv, or whatever) we want our
own on_attach.
To make sure our version is always used, the configuratino is moved to
the after/lsp directory
See https://neovim.io/doc/user/lsp.html#_faq

Q: How to avoid my own lsp/ folder being overridden?
A: Place your configs under "after/lsp/".
   Files in "after/lsp/" are loaded after those in "nvim/lsp/",
   so your settings will take precedence over the defaults provided
   by nvim-lspconfig. See also: after-directory